Conflicts:
	server/views/sources/collection.py
	server/views/sources/source.py

+++ server/views/topics/stories.py
@@ -2,10 +2,12 @@ import logging
 from flask import jsonify, request, Response
 import flask_login
 from multiprocessing import Pool
+from mediacloud.error import MCException
 
 from server import app, cliff, TOOL_API_KEY
 from server.auth import is_user_logged_in
 from server.cache import cache, key_generator
+from server.views import WORD_COUNT_DOWNLOAD_LENGTH
 import server.util.csv as csv
 import server.util.tags as tag_util
 from server.util.request import api_error_handler
@@ -27,7 +29,8 @@ MEDIA_INFO_POOL_SIZE = 15
 def story(topics_id, stories_id):
     if is_user_logged_in():
         local_mc = user_mediacloud_client()
-        story_topic_info = topic_story_list(user_mediacloud_key(), topics_id, stories_id=stories_id)['stories'][0]
+        story_topic_info = topic_story_list(user_mediacloud_key(), topics_id, stories_id=stories_id)
+        story_topic_info = story_topic_info['stories'][0]
         '''
         all_fb_count = []
         more_fb_count = True
@@ -49,18 +52,23 @@ def story(topics_id, stories_id):
     else:
         return jsonify({'status': 'Error', 'message': 'Invalid attempt'})
 
-    story_info = local_mc.story(stories_id)  # add in other fields from regular call
-    for k in story_info.keys():
-        story_topic_info[k] = story_info[k]
-    for tag in story_info['story_tags']:
-        if tag['tag_sets_id'] == tag_util.GEO_TAG_SET:
-            geonames_id = int(tag['tag'][9:])
-            try:
-                tag['geoname'] = _cached_geoname(geonames_id)
-            except Exception as e:
-                # query to CLIFF failed :-( handle it gracefully
-                logger.exception(e)
-                tag['geoname'] = {}
+    try:
+        story_info = local_mc.story(stories_id)  # add in other fields from regular call
+        for k in story_info.keys():
+            story_topic_info[k] = story_info[k]
+        for tag in story_info['story_tags']:
+            if tag['tag_sets_id'] == tag_util.GEO_TAG_SET:
+                geonames_id = int(tag['tag'][9:])
+                try:
+                    tag['geoname'] = _cached_geoname(geonames_id)
+                except Exception as e:
+                    # query to CLIFF failed :-( handle it gracefully
+                    logger.exception(e)
+                    tag['geoname'] = {}
+    except MCException:
+        logger.warning("Story {} wasn't found in a regular story API call, but is it topic {}".format(
+            stories_id, topics_id
+        ))
     return jsonify(story_topic_info)
 
 
@@ -86,7 +94,7 @@ def topic_story_update(stories_id):
 
 @cache.cache_on_arguments(function_key_generator=key_generator)
 def _cached_geoname(geonames_id):
-    return cliff.geonamesLookup(geonames_id)
+    return cliff.geonames_lookup(geonames_id)
 
 
 @app.route('/api/topics/<topics_id>/stories/counts', methods=['GET'])
@@ -145,7 +153,8 @@ def story_words(topics_id, stories_id):
 def story_words_csv(topics_id, stories_id):
     query = add_to_user_query('stories_id:'+stories_id)
     ngram_size = request.args['ngram_size'] if 'ngram_size' in request.args else 1  # default to word count
-    word_counts = topic_ngram_counts(user_mediacloud_key(), topics_id, ngram_size, q=query)
+    word_counts = topic_ngram_counts(user_mediacloud_key(), topics_id, ngram_size, q=query,
+                                     num_words=WORD_COUNT_DOWNLOAD_LENGTH)
     return csv.stream_response(word_counts, WORD_COUNT_DOWNLOAD_COLUMNS,
                                'topic-{}-story-{}-sampled-ngrams-{}-word'.format(topics_id, stories_id, ngram_size))
 
@@ -211,9 +220,9 @@ def stream_story_list_csv(user_key, filename, topics_id, **kwargs):
     merged_args = {
         'snapshots_id': request.args['snapshotId'],
         'timespans_id': request.args['timespanId'],
-        'foci_id': request.args['focusId'],
-        'q': request.args['q'],
-        'sort': request.args['sort'],
+        'foci_id': request.args['focusId'] if 'foci_id' in request.args else None,
+        'q': request.args['q'] if 'q' in request.args else None,
+        'sort': request.args['sort'] if 'sort' in request.args else None,
     }
     params.update(merged_args)
     #
@@ -223,7 +232,7 @@ def stream_story_list_csv(user_key, filename, topics_id, **kwargs):
         del params['fb_data']
     if 'q' in params:
         params['q'] = params['q'] if 'q' not in [None, '', 'null', 'undefined'] else None
-    params['limit'] = 1000  # an arbitrary value to let us page through with big pages
+    params['limit'] = 1000  # an arbitrary value to let us page through with big topics
 
     props = ['stories_id', 'publish_date', 'title', 'url', 'language', 'ap_syndicated',
              'themes','subtopics',
@@ -265,12 +274,17 @@ def stream_story_list_csv(user_key, filename, topics_id, **kwargs):
 def _topic_story_list_by_page_as_csv_row(user_key, topics_id, props, **kwargs):
     yield u','.join(props) + u'\n'  # first send the column names
     link_id = 0
-    for page in _topic_story_page_with_media(user_key, topics_id, link_id, **kwargs):
-        for story in page['stories']:
-            cleaned_row = csv.dict2row(props, story)
+    more_pages = True
+    while more_pages:
+        page = _topic_story_page_with_media(user_key, topics_id, link_id, **kwargs)
+        if 'next' in page['link_ids']:
+            link_id = page['link_ids']['next']
+        else:
+            more_pages = False
+        for s in page['stories']:
+            cleaned_row = csv.dict2row(props, s)
             row_string = u','.join(cleaned_row) + u'\n'
             yield row_string
-        link_id += 1
 
 
 def _media_info_worker(info):
@@ -279,27 +293,31 @@ def _media_info_worker(info):
 
 # generator you can use to do something for each page of story results
 def _topic_story_page_with_media(user_key, topics_id, link_id, **kwargs):
+    add_media_fields = False  # switch for including all the media metadata in each row (ie. story)
+
     story_page = topic_story_list_by_page(user_key, topics_id, link_id=link_id, **kwargs)
 
     story_ids = [str(s['stories_id']) for s in story_page['stories']]
     stories_with_tags = story_list(user_key, 'stories_id:(' + " ".join(story_ids) + ")", kwargs['limit'])
 
     # build a media lookup table in parallel so it is faster
-    pool = Pool(processes=MEDIA_INFO_POOL_SIZE)
-    jobs = [{'user_key': user_key, 'media_id': s['media_id']} for s in story_page['stories']]
-    job_results = pool.map(_media_info_worker, jobs)  # blocks until they are all done
-    media_lookup = {j['media_id']: j for j in job_results}
-    pool.terminate()
+    if add_media_fields:
+        pool = Pool(processes=MEDIA_INFO_POOL_SIZE)
+        jobs = [{'user_key': user_key, 'media_id': s['media_id']} for s in story_page['stories']]
+        job_results = pool.map(_media_info_worker, jobs)  # blocks until they are all done
+        media_lookup = {j['media_id']: j for j in job_results}
+        pool.terminate()
 
     # update story info for each story in the page, put it into the [stories] field, send updated page with stories back
     for s in story_page['stories']:
 
         # add in media metadata to the story (from page-level cache built earlier)
-        media = media_lookup[s['media_id']]
+        if add_media_fields:
+            media = media_lookup[s['media_id']]
 
-        # add in foci/subtopic names
-        for k, v in media['metadata'].iteritems():
-             s[u'media_{}'.format(k)] = v['label'] if v is not None else None
+            # add in foci/subtopic names
+            for k, v in media['metadata'].iteritems():
+                s[u'media_{}'.format(k)] = v['label'] if v is not None else None
 
         # build lookup for id => story for all stories in stories with tags (non topic results)
         for st in stories_with_tags:
@@ -315,9 +333,5 @@ def _topic_story_page_with_media(user_key, topics_id, link_id, **kwargs):
                 if tag_util.NYT_LABELER_1_0_0_TAG_ID in story_tag_ids:
                     story_tag_ids = [t['tag'] for t in s['story_tags'] if t['tag_sets_id'] == tag_util.NYT_LABELS_TAG_SET_ID]
                     s['themes'] = ", ".join(story_tag_ids)
-                # s is updated
-                # how do I add s back into story_page?
 
-
-    yield story_page # need links too
-
+    return story_page  # need links too
